Add SQL database setup
